minor bug fix
This problem only affects those who may have used accuracy check or in any other way called model.eval() in the training loop. Our original code did not do this, so there was no need for a model.train() in the triaining loop, but we are adding it now to be on the safer side.